split the esp32-idf 4.x vs 5.x rmt driver. 5.x just redirects to 4.x
