+++ src/namespace.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#ifndef FASTLED_NAMESPACE_BEGIN
+/// Start of the FastLED namespace
+#define FASTLED_NAMESPACE_BEGIN
+/// End of the FastLED namespace
+#define FASTLED_NAMESPACE_END
+/// "Using" directive for the namespace
+#define FASTLED_USING_NAMESPACE
+#endif
