+++ src/platforms/esp/32/idf5_clockless_rmt_esp32.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#warning "Using the legacy RMT driver. This will crash your system if any other code initializes the new RMT driver."
+#include "idf4_clockless_rmt_esp32.h"
